requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
airflow-provider-duckdb
pyarrow
dbt-duckdb
dbt-osmosis
openai